+++ packages/standard-minifiers/package.js
@@ -5,22 +5,8 @@ Package.describe({
   documentation: 'README.md'
 });
 
-Package.registerBuildPlugin({
-  name: "minifyStd",
-  use: [
-    'minifiers'
-  ],
-  npmDependencies: {
-    "source-map": "0.4.2"
-  },
-  sources: [
-    'plugin/minify-js.js',
-    'plugin/minify-css.js'
-  ]
-});
-
 Package.onUse(function(api) {
-  api.use('isobuild:minifier-plugin@1.0.0');
+  api.imply(['standard-minifiers-css','standard-minifiers-js']);
 });
 
 Package.onTest(function(api) {